fixed computationDAO, computationStepDAO and reorganized manager, making it more dynamic. Almost finished recent Nets visualization

+++ src/main/java/it/petrinet/model/PetriNet.java
@@ -3,12 +3,12 @@ package it.petrinet.model;
 public class PetriNet {
     private String netName;
     private String creatorId;
-    private int creationDate;
+    private Long creationDate;
     private String XML_PATH;
     private String image_PATH;
     private boolean isReady;
 
-    public PetriNet(String netName, String creatorId, int creationDate, String XML_PATH, String imagePATH, boolean isReady){
+    public PetriNet(String netName, String creatorId, Long creationDate, String XML_PATH, String imagePATH, boolean isReady){
         this.netName = netName;
         this.creatorId = creatorId;
         this.creationDate = creationDate;
@@ -26,7 +26,7 @@ public class PetriNet {
         return creatorId;
     }
 
-    public int getCreationDate() {
+    public Long getCreationDate() {
         return creationDate;
     }
